Recriando projeto Web. Adicionando documentação na API. Ajustando database para  sqlite.
