+++ Restaurante.Clientes.API/appsettings.json
@@ -1,10 +1,13 @@
 {
   "Logging": {
-      "LogLevel": {
+    "LogLevel": {
       "Default": "Information",
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
-      }
-    },
-"AllowedHosts": "*"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "Default": "Data Source=Restaurante.db"
+  }
 }
