--- Restaurante.Clientes.WebSite/ClientApp/.editorconfig
@@ -1,16 +1,0 @@
-# Editor configuration, see http://editorconfig.org
-root = true
-
-[*]
-charset = utf-8
-indent_style = space
-indent_size = 2
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[*.md]
-max_line_length = off
-trim_trailing_whitespace = false
-
-[*.{razor,cshtml}]
-charset = utf-8-bom
