--- Restaurante.Clientes.WebSite/ClientApp/src/app/app.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html'
-})
-export class AppComponent {
-  title = 'app';
-}
