--- Restaurante.Clientes.WebSite/ClientApp/src/app/nav-menu/nav-menu.component.ts
@@ -1,18 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-nav-menu',
-  templateUrl: './nav-menu.component.html',
-  styleUrls: ['./nav-menu.component.css']
-})
-export class NavMenuComponent {
-  isExpanded = false;
-
-  collapse() {
-    this.isExpanded = false;
-  }
-
-  toggle() {
-    this.isExpanded = !this.isExpanded;
-  }
-}
