--- Restaurante.Clientes.WebSite/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using Restaurante.Clientes.WebSite
-@namespace Restaurante.Clientes.WebSite.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
