+++ Restaurante.Clientes.sln
@@ -11,7 +11,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Restaurante.Clientes.Applic
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Restaurante.Clientes.Infra", "Restaurante.Infra\Restaurante.Clientes.Infra.csproj", "{A8C32B58-5955-4114-A289-C15262A2D151}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Restaurante.Clientes.WebSite", "Restaurante.Clientes.WebSite\Restaurante.Clientes.WebSite.csproj", "{FDB141AC-B27B-4E33-90C5-3C6BCFBA7769}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Restaurante.Clientes.API", "Restaurante.Clientes.API\Restaurante.Clientes.API.csproj", "{53D02245-F820-4448-A2C0-8794AC7CC4D0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -35,10 +35,10 @@ Global
 		{A8C32B58-5955-4114-A289-C15262A2D151}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A8C32B58-5955-4114-A289-C15262A2D151}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A8C32B58-5955-4114-A289-C15262A2D151}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FDB141AC-B27B-4E33-90C5-3C6BCFBA7769}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FDB141AC-B27B-4E33-90C5-3C6BCFBA7769}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FDB141AC-B27B-4E33-90C5-3C6BCFBA7769}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FDB141AC-B27B-4E33-90C5-3C6BCFBA7769}.Release|Any CPU.Build.0 = Release|Any CPU
+		{53D02245-F820-4448-A2C0-8794AC7CC4D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{53D02245-F820-4448-A2C0-8794AC7CC4D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{53D02245-F820-4448-A2C0-8794AC7CC4D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{53D02245-F820-4448-A2C0-8794AC7CC4D0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
